OVN Support
Pacemaker support for OVN DB servers

This patch enables ovn database servers to be started in active/backup
mode in the pacemaker cluster.

A virtual IP resource will be co-hosted with the active database server
so that, the services (like ovn-controllers and neutron OVN ML2
mechanism driver) which depends on ovn database servers (will always connect
to a single IP address on which the active database server listens on.

ovn northd is managed by the pacemaker and it is colocated on the same
node where the virtual IP resides.

Also, when HA is enabled for ovn database servers, haproxy does not
configure the ovn database servers in its configuration.

Co-authored:by: Numan Siddique <nusiddiq@redhat.com>
Change-Id: I9dc366002ef5919339961e5deebbf8aa815c73db
Closes-bug: #1670564